bug corrections
