--- lab02/analisamem/teste/transplog.out
@@ -1,127 +1,0 @@
-I 1 616722.227274644
-E 0 2 0 0.000013122 140727060730464 8
-E 0 3 0 0.000014357 140727060730472 8
-E 0 4 0 0.000014949 140727060730480 8
-E 0 5 0 0.000015410 140727060730488 8
-E 0 6 0 0.000015860 140727060730496 8
-E 0 7 0 0.000016385 140727060730504 8
-E 0 8 0 0.000016915 140727060730512 8
-E 0 9 0 0.000017377 140727060730520 8
-E 0 10 0 0.000017806 140727060730528 8
-E 0 11 0 0.000018375 140727060730536 8
-E 0 12 0 0.000018833 140727060730544 8
-E 0 13 0 0.000019302 140727060730552 8
-E 0 14 0 0.000019796 140727060730560 8
-E 0 15 0 0.000020253 140727060730568 8
-E 0 16 0 0.000020681 140727060730576 8
-E 0 17 0 0.000021113 140727060730584 8
-E 0 18 0 0.000021594 140727060730592 8
-E 0 19 0 0.000022102 140727060730600 8
-E 0 20 0 0.000022570 140727060730608 8
-E 0 21 0 0.000023001 140727060730616 8
-E 0 22 0 0.000023531 140727060730624 8
-E 0 23 0 0.000023970 140727060730632 8
-E 0 24 0 0.000024458 140727060730640 8
-E 0 25 0 0.000024893 140727060730648 8
-E 0 26 0 0.000025336 140727060730656 8
-E 0 27 0 0.000026115 140727060730464 8
-E 0 28 0 0.000026582 140727060730472 8
-E 0 29 0 0.000027014 140727060730480 8
-E 0 30 0 0.000027456 140727060730488 8
-E 0 31 0 0.000027926 140727060730504 8
-E 0 32 0 0.000028399 140727060730512 8
-E 0 33 0 0.000028888 140727060730520 8
-E 0 34 0 0.000029380 140727060730528 8
-E 0 35 0 0.000029834 140727060730544 8
-E 0 36 0 0.000030242 140727060730552 8
-E 0 37 0 0.000030737 140727060730560 8
-E 0 38 0 0.000031157 140727060730568 8
-E 0 39 0 0.000031622 140727060730584 8
-E 0 40 0 0.000032102 140727060730592 8
-E 0 41 0 0.000032616 140727060730600 8
-E 0 42 0 0.000033055 140727060730608 8
-E 0 43 0 0.000033521 140727060730624 8
-E 0 44 0 0.000033973 140727060730632 8
-E 0 45 0 0.000034452 140727060730640 8
-E 0 46 0 0.000034915 140727060730648 8
-L 1 47 0 0.000035466 140727060730464 8
-L 1 48 0 0.000036149 140727060730472 8
-L 1 49 0 0.000036646 140727060730480 8
-L 1 50 0 0.000037095 140727060730488 8
-L 1 51 0 0.000037622 140727060730504 8
-L 1 52 0 0.000038132 140727060730512 8
-L 1 53 0 0.000038639 140727060730520 8
-L 1 54 0 0.000039262 140727060730528 8
-L 1 55 0 0.000039807 140727060730544 8
-L 1 56 0 0.000040347 140727060730552 8
-L 1 57 0 0.000040774 140727060730560 8
-L 1 58 0 0.000041228 140727060730568 8
-L 1 59 0 0.000041686 140727060730584 8
-L 1 60 0 0.000042178 140727060730592 8
-L 1 61 0 0.000042683 140727060730600 8
-L 1 62 0 0.000043208 140727060730608 8
-L 1 63 0 0.000043648 140727060730624 8
-L 1 64 0 0.000044188 140727060730632 8
-L 1 65 0 0.000044672 140727060730640 8
-L 1 66 0 0.000045150 140727060730648 8
-E 1 67 0 0.000045836 140727060730472 8
-E 1 68 0 0.000046419 140727060730504 8
-E 1 69 0 0.000046930 140727060730480 8
-E 1 70 0 0.000047452 140727060730544 8
-E 1 71 0 0.000047938 140727060730488 8
-E 1 72 0 0.000048440 140727060730584 8
-E 1 73 0 0.000048963 140727060730496 8
-E 1 74 0 0.000049414 140727060730624 8
-E 1 75 0 0.000049908 140727060730520 8
-E 1 76 0 0.000050387 140727060730552 8
-E 1 77 0 0.000050851 140727060730528 8
-E 1 78 0 0.000051274 140727060730592 8
-E 1 79 0 0.000051724 140727060730536 8
-E 1 80 0 0.000052180 140727060730632 8
-E 1 81 0 0.000052653 140727060730568 8
-E 1 82 0 0.000053082 140727060730600 8
-E 1 83 0 0.000053568 140727060730576 8
-E 1 84 0 0.000053984 140727060730640 8
-E 1 85 0 0.000054516 140727060730616 8
-E 1 86 0 0.000054962 140727060730648 8
-L 2 87 0 0.000055508 140727060730464 8
-L 2 88 0 0.000056014 140727060730472 8
-L 2 89 0 0.000056517 140727060730480 8
-L 2 90 0 0.000056977 140727060730488 8
-L 2 91 0 0.000057447 140727060730496 8
-L 2 92 0 0.000057934 140727060730504 8
-L 2 93 0 0.000058404 140727060730512 8
-L 2 94 0 0.000058864 140727060730520 8
-L 2 95 0 0.000059333 140727060730528 8
-L 2 96 0 0.000059850 140727060730536 8
-L 2 97 0 0.000060356 140727060730544 8
-L 2 98 0 0.000060849 140727060730552 8
-L 2 99 0 0.000061355 140727060730560 8
-L 2 100 0 0.000061804 140727060730568 8
-L 2 101 0 0.000062441 140727060730576 8
-L 2 102 0 0.000062956 140727060730584 8
-L 2 103 0 0.000063419 140727060730592 8
-L 2 104 0 0.000063885 140727060730600 8
-L 2 105 0 0.000064335 140727060730608 8
-L 2 106 0 0.000064864 140727060730616 8
-L 2 107 0 0.000088798 140727060730464 8
-L 2 108 0 0.000090069 140727060730472 8
-L 2 109 0 0.000090802 140727060730480 8
-L 2 110 0 0.000091759 140727060730488 8
-L 2 111 0 0.000092613 140727060730496 8
-L 2 112 0 0.000096728 140727060730504 8
-L 2 113 0 0.000097702 140727060730512 8
-L 2 114 0 0.000098560 140727060730520 8
-L 2 115 0 0.000099383 140727060730528 8
-L 2 116 0 0.000100217 140727060730536 8
-L 2 117 0 0.000102994 140727060730544 8
-L 2 118 0 0.000104047 140727060730552 8
-L 2 119 0 0.000104811 140727060730560 8
-L 2 120 0 0.000105534 140727060730568 8
-L 2 121 0 0.000106375 140727060730576 8
-L 2 122 0 0.000108582 140727060730584 8
-L 2 123 0 0.000109414 140727060730592 8
-L 2 124 0 0.000110264 140727060730600 8
-L 2 125 0 0.000111333 140727060730608 8
-L 2 126 0 0.000112026 140727060730616 8
-F 127 616722.227388850 0.000114206
