+++ lab02/include/fatorial.hpp
@@ -1,5 +1,10 @@
+#ifndef TP2_ED_FATORIAL_H
+#define TP2_ED_FATORIAL_H
+
 #include <iostream>
 
-int fatorial_iterativo (int n);
-int fatorial_recursivo (int n);
-int fatorial_recursivo_mod (int n);
+long long int fatorial_iterativo (int n);
+long long int fatorial_recursivo (int n);
+long long int fatorial_recursivo_mod (int n);
+
+#endif
