+++ lab02/include/fibonacci.hpp
@@ -1,4 +1,8 @@
+#ifndef TP2_ED_FIBONACCI_H
+#define TP2_ED_FIBONACCI_H
 
 int fibonacci_iterativo (int n);
 int fibonacci_recursivo (int n);
 int fibonacci_recursivo_mod (int n);
+
+#endif
