--- lab02/include/test_function.hpp
@@ -1,18 +1,0 @@
-#include <iostream>
-
-#include "./fatorial.hpp"
-#include "./fibonacci.hpp"
-#include "./time_count.hpp"
-
-// IDEIA: FAZER UMA FUNCAO QUE JÁ MEDE OS TEMPOS, DIMINUINDO O TAMANHO DO MAIN
-
-
-// void test_function(function nome_funcao, int n) { 
-//     start_count();
-//     std::cout << nome_funcao << std::endl;
-//     for (int i = 0; i < n; i++) {
-//         std::cout << "n = " << i << std::endl;
-//         std::cout << nome_funcao << nome_funcao(i) << std::endl;
-//     }
-//     end_count();
-// }
