+++ lab02/src/fatorial.cpp
@@ -3,20 +3,20 @@
 
 #include "../include/fatorial.hpp"
 
-int fatorial_iterativo (int n) {
-    int total = 1;
+long long int fatorial_iterativo (int n) {
+    long long int total = 1;
     for (; n > 1; n--) {
         total *= n;
     }
     return total;
 }
 
-int fatorial_recursivo (int n) {
+long long int fatorial_recursivo (int n) {
     if (n == 0) return 1;
     else return n * fatorial_recursivo(n-1);
 }
 
-int fatorial_recursivo_mod (int n) {
+long long int fatorial_recursivo_mod (int n) {
     if (n == 0) return 1;
     else {
         double result = 0.0;
